refactor(agent-toolkit): consolidate workflow write tools into manage_workflows
Merge set_workflow_active_state and delete_workflow into a single manage_workflows
tool with action enum (activate/deactivate/delete), keeping list_workflows as a
separate READ tool for proper read/write separation.
