+++ packages/agent-toolkit/src/core/tools/platform-api-tools/workflows-tools/index.ts
@@ -1,3 +1,2 @@
 export * from './list-workflows/list-workflows-tool';
-export * from './set-workflow-active-state/set-workflow-active-state-tool';
-export * from './delete-workflow/delete-workflow-tool';
+export * from './manage-workflows/manage-workflows-tool';
