+++ packages/agent-toolkit/src/core/tools/platform-api-tools/workflows-tools/manage-workflows/workflow-mutations.graphql.dev.ts
@@ -15,3 +15,11 @@ export const deactivateLiveWorkflowMutation = gql`
     }
   }
 `;
+
+export const deleteLiveWorkflowMutation = gql`
+  mutation deleteLiveWorkflow($id: ID!) {
+    delete_live_workflow(id: $id) {
+      is_success
+    }
+  }
+`;
